Update global_cv_analysis.py
Added SSP3 population scenario (under "poverty" - just because it used a .tif and not a .zip. In the future would be great to make this script allow different population scenarios or to take population as a tif instead of the very specific structure of the Landscan zip files. But this hack will have to do for now.